gridRenderer became screenRenderer. We can now render things on a per scene basis.
