+++ src/scenes/menu.js
@@ -1,10 +1,13 @@
 import Phaser from "phaser";
 import logoImg from "../assets/Sprites/logo.png";
+import BrownRock from "../assets/Sprites/brown_rock.png"
+import GreenRock from "../assets/Sprites/green_rock.png"
 
 export default class Menu extends Phaser.Scene {
   preload() {
-    // this.load.image("logo", logoImg);
-    console.log(logoImg);
+    this.load.image("logo", logoImg);
+    this.load.image("brownRock",BrownRock);
+    this.load.image("greenRock",GreenRock)
   }
 
   create() {
